fix 4/28 류예나  게시물 전체조회, 카테고리 조회 구현

+++ src/main/java/HeartGuard/Board/model/dto/ReplyDto.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class ReplyDto {
-    private int rno;
+    private long rno;
     private String rcontent;
     private LocalDateTime rdate;
     private int uno;
